Release v0.1.1 (#33)
* Fixed ToC for `mocking` checker
([#29](https://github.com/databrickslabs/pylint-plugin/issues/29)). In
this release, we have made significant improvements to the `mocking`
checker in our open-source library, which is specifically designed for
identifying common mistakes and issues in Spark code written in Python.
We have added two new rules, `R8918: explicit-dependency-required` and
`R8919: obscure-mock`, which respectively check for the requirement of
explicit dependencies and the use of obscure mocks. This update enhances
the accuracy and quality of the code review process for Spark code
written in Python, ensuring that the code meets the highest standards of
quality and reliability. Additionally, we have fixed an issue in the
Table of Contents (ToC) for the `mocking` checker, making it easier for
users to navigate and utilize this feature.
* Fixed project urls for PyPI
([#32](https://github.com/databrickslabs/pylint-plugin/issues/32)). In
this release, the `pyproject.toml` configuration file for the Python
project has been updated to fix project URLs for PyPI. Specifically, the
URLs for the `Issues` and `Source` fields have been changed from
"<https://github.com/databrickslabs/pylint/issues>" and
"<https://github.com/databrickslabs/pylint>" to
"<https://github.com/databrickslabs/pylint-plugin/issues>" and
"<https://github.com/databrickslabs/pylint-plugin>", respectively. This
modification ensures that users can correctly report issues and access
the source code for the pylint-plugin project on GitHub. The rest of the
file remains unchanged. This enhancement provides a seamless experience
for users who want to contribute to the project or seek support for any
issues they encounter.
* Updated README.md to fix pip install command
([#30](https://github.com/databrickslabs/pylint-plugin/issues/30)). In
this release, we have updated the installation command for the project
to fix the pip install process. Previously, users installed the project
using the command "pip install pylint-plugin-for-databricks", but this
has been changed to "pip install databricks-labs-pylint". This update is
part of issue
[#30](https://github.com/databrickslabs/pylint-plugin/issues/30) and is
intended to improve the installation experience for users. It is
important to note that no new methods have been added and no existing
functionality has been changed; only the installation command has been
updated. Software engineers adopting this project should use the new
command to install it.